requirements.txt
mlflow==2.14.1
scikit-learn==1.4.2
pandas==2.1.4
numpy==1.26.4
joblib==1.3.2
pyyaml==6.0.1
pytest==8.2.0
flake8==7.0.0
